Utilize ThreadPoolExecutor to run SQL queries concurrently
Refactored the SQL querying process in main_app.py to make it run concurrently using the ThreadPoolExecutor from the concurrent.futures module. In the earlier version, all database queries were executed sequentially which resulted in the overall execution time of the script actually being the sum of the execution time of each query. By utilizing a ThreadPoolExecutor to run execute_query_and_get_count in parallel for each query, the querying process becomes concurrent, thus greatly reducing the overall execution time of the script especially when the quantity of the data being processed is large. Also refactored the method of handling date formatting in the DataFrame.