Added pandas display options for better data readability
Custom display options for Pandas dataframe have been added to aid in better read and understanding of large datasets. Max columns and max rows were set to None so that all columns and rows will be displayed, and display width was set to 1000 to avoid truncation. Changes have been made to make data processing and review more efficient and readable.