Expand RankedRecords SQL query and add user filter
Extended the RankedRecords SQL query to retrieve the top 10 records, as opposed to the sole first one. This is due to errors occuring with multiple login-logout events.

Furthermore, added a 'fix_check_online_user' function to identify unique user sessions in the case of potential mismatches. This ensures correct user status identification.