Update date parameters and color codes
Adjusted the date parameters to retrieve data from 11 years ago instead of 5, for a more comprehensive analysis and to cover periods of significant business events. Also, changed the color codes used to represent customer count. Instead of orange, "#F25E49" is now used for a clearer distinction. Commented out 'l' and 'a01' from the array of files to focus on 'a0' for this iteration.