Refactor 'run' function in write.py, disable unused parameters
In response to changes in data requirements leading to exceptions, unused parameters 'customers' and 'customers_month' were commented out in 'run' function in write.py. This step reduces potential errors and optimizes the overall function performance by not assigning unused variables. In the future, when data retrieval issues are resolved, these parameters can be uncommented and used accordingly. Added a condition to prevent unnecessary calendar creation, which helps to optimize the process further.