Remove lato_users data fetch for performance improvement
Lato_users' data fetch along with the subsequent processing has been commented out in this change due to its redundancy and the significant lag it was creating. This particular data is not required for the current operation of the app, hence its removal. Now, the application will focus on only elounda_users data fetching and processing, which streamlines execution and improves overall performance.