Adjust data visualization and image overlay placement
Changes in plot.py include commenting out the plotting and fill_between function for Y_all data, which improves graph clarity without the 'grey' overlay. To handle negative numbers, label is adjusted to skip 'EUR' for negative b-values. Also, an adjustment was made to the position of the image overlay from (343, 4000) to (1250, 2300) for better location accuracy.