Updated font and image positioning in write.py
This commit includes several changes in write.py aimed at enhancing readability and presentation. A change in font from "Futura.ttc" to "Avenir Next.ttc" and "DIN Condensed Bold.ttf" for title and numbers respectively was done to improve visibility and aesthetics.

The location of image overlays was modified to increase visualization quality. For status_users_lato, data was removed due to its redundancy, thereby decluttering the output image. The potential mishandling of negative numbers was also checked by adjusting labels to skip 'EUR' for negative figures.

Additionally, the positioning of the calendar image was adjusted for better visibility. Lastly, the commit introduces use of the 'a01' flag for the inclusion of daily plot operations.