Shorten years difference from 11 to 5 in main_app.py
This commit modifies the "params_1" and "params_3" dictionaries in main_app.py to reduce the range of years considered from 11 to 5. This change was necessary to keep the data analysis within a more recent context and improve its relevance. Same change was applied in write.py. Alteration of various elements' locations on image also happened in write.py due to the given requirement of adjusting workspace.