Change font family and representation of currency in plot.py
Modified font family from monospace to PoiretOne and currency representation from '\nEUR' to '€' in plot.py. The font change was implemented to match the overall aesthetic of the project. The currency representation has been streamlined to make the displayed information cleaner and more readable. Also, optimized some for-loop operations and added font properties.