Refactor code to display customers_month data on image editable
The code was adjusted in the 'write.py' file to display the 'customers_month' data on the image editable. The adjustments include uncommenting some sections related to 'customers_month' which allow these values to be included in the rendered image. Furthermore, certain coordinates were changed to fit the 'customer_month' data within the allocated space. This alteration was required to provide more comprehensive year-on-year customer data information through these images.