Add vpn error handling functionality
Added a function call to 'sql_connect.open_vpn(failed)' when VPN connection is offline in main_app.py. It aids in managing the failed VPN connection attempts and provides a better response to such situations. This change helps to debug issues with VPN connectivity and institutes a more reliable application behaviour in case of such network failures.