Update Pillow version in requirements.txt
Updated the version requirement of Pillow in requirements.txt. Instead of strictly using Pillow version 10.0.0, the code is now compatible with any version of Pillow starting from 10.0.0 and above. This change allows for greater flexibility with dependency resolutions and can avoid potential collisions with other packages that may need a newer version of Pillow. Also, it increases the possibility of benefiting from future updates of the Pillow library.