"Added DEH self-metering to scheduled activities, Simplified draw text condition
This commit adds "DEH Self Metering" activity to the scheduled activities dictionary, allowing tracking for this function. The visualization of these activities has also been enhanced by simplifying the conditions for the draw text function. Now, the text appears not only when the event is in the current month (with no specific month set) but also when the event is today or tomorrow, thus improving readability and user interaction."